requirements.txt
apiview==1.3.25
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.8.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
future==0.18.2
idna==3.3
inflection==0.5.1
iso8601==1.0.2
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
model-mommy==2.0.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
serial==0.0.97
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
values==2020.12.3
xlrd==2.0.1
xlwt==1.3.0
